+++ src/app/result/[postId]/_components/VoteBlock/VoteBlockForm.tsx
